Replace Tesla with Tesla Command integration
