+++ vehicle/tesla-command/helper.go
